requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
psycopg==3.1.17
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
